+++ apis/nucleus/src/components/listbox/ListBoxInline.jsx
@@ -19,7 +19,15 @@ import useAppSelections from '../../hooks/useAppSelections';
 import showToolbarDetached from './interactions/listbox-show-toolbar-detached';
 import getListboxActionProps from './interactions/listbox-get-action-props';
 import createSelectionState from './hooks/selections/selectionState';
-import { CELL_PADDING_LEFT, ICON_WIDTH, ICON_PADDING, BUTTON_ICON_WIDTH, HEADER_PADDING_RIGHT } from './constants';
+import {
+  CELL_PADDING_LEFT,
+  ICON_WIDTH,
+  ICON_PADDING,
+  BUTTON_ICON_WIDTH,
+  HEADER_PADDING_RIGHT,
+  DENSE_ROW_HEIGHT,
+  SCROLL_BAR_WIDTH,
+} from './constants';
 import useTempKeyboard from './components/useTempKeyboard';
 import ListBoxError from './components/ListBoxError';
 import useRect from '../../hooks/useRect';
@@ -220,7 +228,7 @@ function ListBoxInline({ options, layout }) {
     }
   }, [searchContainer && searchContainer.current, showSearch, search, focusSearch]);
 
-  const { wildCardSearch, searchEnabled, layoutOptions = {} } = layout ?? {};
+  const { wildCardSearch, searchEnabled, autoConfirm = false, layoutOptions = {} } = layout ?? {};
   const showSearchIcon = searchEnabled !== false && search === 'toggle';
   const isLocked = layout?.qListObject?.qDimensionInfo?.qLocked === true;
   const showSearchOrLockIcon = isLocked || showSearchIcon;
@@ -262,10 +270,6 @@ function ListBoxInline({ options, layout }) {
   const showSearchToggle = search === 'toggle' && showSearch;
   const searchVisible = (search === true || showSearchToggle) && !selectDisabled() && searchEnabled !== false;
   const dense = layoutOptions.dense ?? false;
-  const searchHeight = dense ? 27 : 40;
-  const extraheight = dense ? 39 : 49;
-  const searchAddHeight = searchVisible ? searchHeight : 0;
-  const minHeight = showToolbarWithTitle ? 49 + searchAddHeight + extraheight : 0;
   const headerHeight = 32;
 
   const onShowSearch = () => {
@@ -282,8 +286,8 @@ function ListBoxInline({ options, layout }) {
     }
   };
 
-  const getActionToolbarProps = (isDetached) =>
-    getListboxActionProps({
+  const getActionToolbarProps = (isDetached) => ({
+    ...getListboxActionProps({
       isDetached: isPopover ? false : isDetached,
       showToolbar,
       containerRef,
@@ -291,7 +295,9 @@ function ListBoxInline({ options, layout }) {
       listboxSelectionToolbarItems,
       selections,
       keyboard,
-    });
+    }),
+    autoConfirm,
+  });
 
   const shouldAutoFocus = searchVisible && search === 'toggle';
 
@@ -339,7 +345,7 @@ function ListBoxInline({ options, layout }) {
         direction="column"
         gap={0}
         containerPadding={containerPadding}
-        style={{ height: '100%', minHeight: `${minHeight}px`, flexFlow: 'column nowrap' }}
+        style={{ height: '100%', flexFlow: 'column nowrap' }}
         onKeyDown={handleKeyDown}
         onMouseEnter={handleOnMouseEnter}
         onMouseLeave={handleOnMouseLeave}
@@ -348,6 +354,7 @@ function ListBoxInline({ options, layout }) {
           containerRectRef(el);
         }}
         isGridMode={isGridMode}
+        aria-label={keyboard.active ? translator.get('Listbox.ScreenReaderInstructions') : ''}
       >
         {showToolbarWithTitle && (
           <Grid
@@ -395,12 +402,11 @@ function ListBoxInline({ options, layout }) {
           item
           container
           direction="column"
-          style={{ height: '100%', minHeight: '50px' }}
+          style={{ height: '100%', minHeight: DENSE_ROW_HEIGHT + SCROLL_BAR_WIDTH }}
           role="region"
           aria-label={translator.get('Listbox.ResultFilterLabel')}
         >
           <Grid item ref={searchContainerRef}>
-            <div className={classes.screenReaderOnly}>{translator.get('Listbox.Search.ScreenReaderInstructions')}</div>
             <ListBoxSearch
               selections={selections}
               selectionState={selectionState}
@@ -417,7 +423,6 @@ function ListBoxInline({ options, layout }) {
             />
           </Grid>
           <Grid item xs className={classes.listboxWrapper}>
-            <div className={classes.screenReaderOnly}>{translator.get('Listbox.ScreenReaderInstructions')}</div>
             {isInvalid ? (
               <ListBoxError text={errorText} />
             ) : (