+++ apis/nucleus/src/sheet.js
@@ -92,6 +92,9 @@ export default function sheet({ model, halo, initialError, onDestroy = async () 
         if (mountedReference) {
           throw new Error('Already mounted');
         }
+        if (!(element instanceof HTMLElement)) {
+          throw new Error('Provided element is not a proper HTMLElement');
+        }
         mountedReference = element;
         [unmountSheet, sheetRef] = glueSheet({
           halo,