+++ apis/stardust/types/index.d.ts
@@ -182,7 +182,7 @@ export function useRenderState(): stardust.RenderState;
 /**
  * Gets an event emitter instance for the visualization.
  */
-export function useEmitter(): Emitter;
+export function useEmitter(): stardust.Emitter;
 
 /**
  * Gets the desired keyboard settings and status to applied when rendering the visualization.
@@ -395,11 +395,16 @@ declare namespace stardust {
         destroy(): void;
 
         /**
-         * Converts the visualization to a different registered type
+         * Converts the visualization to a different registered type.
+         * 
+         * Will update properties if permissions allow, else will patch (can be forced with forcePatch parameter)
+         * 
+         * Not all chart types are compatible, similar structures are required.
          * @param newType Which registered type to convert to.
-         * @param forceUpdate Whether to run setProperties or not, defaults to true.
+         * @param forceUpdate Whether to apply the change or not, else simply returns the resulting properties, defaults to true.
+         * @param forcePatch Whether to always patch the change instead of making a permanent change
          */
-        convertTo(newType: string, forceUpdate?: boolean): Promise<object>;
+        convertTo(newType: string, forceUpdate?: boolean, forcePatch?: boolean): Promise<object>;
 
         /**
          * Listens to custom events from inside the visualization. See useEmitter
@@ -504,6 +509,16 @@ declare namespace stardust {
 
     }
 
+    /**
+     * An object literal containing meta information about the plugin and a function containing the plugin implementation.
+     */
+    interface Plugin {
+        info: {
+            name: string;
+        };
+        fn: ()=>void;
+    }
+
     type Field = string | EngineAPI.INxDimension | EngineAPI.INxMeasure | stardust.LibraryField;
 
     /**
@@ -534,16 +549,6 @@ declare namespace stardust {
     interface LibraryField {
         qLibraryId: string;
         type: "dimension" | "measure";
-    }
-
-    /**
-     * An object literal containing meta information about the plugin and a function containing the plugin implementation.
-     */
-    interface Plugin {
-        info: {
-            name: string;
-        };
-        fn: ()=>void;
     }
 
     interface LoadType {
@@ -653,6 +658,11 @@ declare namespace stardust {
     interface RenderState {
         pending: any;
         restore: any;
+    }
+
+    class Emitter {
+        constructor();
+
     }
 
     interface Keyboard {