+++ apis/enigma-mocker/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.5.0](https://github.com/qlik-oss/nebula.js/compare/v4.5.0-alpha.0...v4.5.0) (2023-10-03)
+
+**Note:** Version bump only for package @nebula.js/enigma-mocker
+
+# [4.5.0-alpha.0](https://github.com/qlik-oss/nebula.js/compare/v4.4.0...v4.5.0-alpha.0) (2023-08-31)
+
+### Bug Fixes
+
+- unresolving promises for missing types ([#1340](https://github.com/qlik-oss/nebula.js/issues/1340)) ([de38cbe](https://github.com/qlik-oss/nebula.js/commit/de38cbedf69fd13340834a902da17efc23c055f9))
+
+# [4.4.0](https://github.com/qlik-oss/nebula.js/compare/v4.3.0...v4.4.0) (2023-08-14)
+
+### Bug Fixes
+
+- race condition for mocked charts ([#1329](https://github.com/qlik-oss/nebula.js/issues/1329)) ([6780ba5](https://github.com/qlik-oss/nebula.js/commit/6780ba5c35f2f58d06c2c753068724903ef8bef4))
+
+# [4.3.0](https://github.com/qlik-oss/nebula.js/compare/v4.2.4...v4.3.0) (2023-08-03)
+
+**Note:** Version bump only for package @nebula.js/enigma-mocker
+
+## [4.2.4](https://github.com/qlik-oss/nebula.js/compare/v4.2.3...v4.2.4) (2023-07-04)
+
+**Note:** Version bump only for package @nebula.js/enigma-mocker
+
+## [4.2.3](https://github.com/qlik-oss/nebula.js/compare/v4.2.2...v4.2.3) (2023-06-28)
+
+**Note:** Version bump only for package @nebula.js/enigma-mocker
+
+## [4.2.2](https://github.com/qlik-oss/nebula.js/compare/v4.2.1...v4.2.2) (2023-06-28)
+
+**Note:** Version bump only for package @nebula.js/enigma-mocker
+
 ## [4.2.1](https://github.com/qlik-oss/nebula.js/compare/v4.2.0...v4.2.1) (2023-06-09)
 
 **Note:** Version bump only for package @nebula.js/enigma-mocker